+++ src/main/java/woowacourse/shoppingcart/auth/application/AuthService.java
@@ -5,8 +5,8 @@ import org.springframework.transaction.annotation.Transactional;
 import woowacourse.shoppingcart.auth.dto.LoginRequest;
 import woowacourse.shoppingcart.customer.application.CustomerService;
 import woowacourse.shoppingcart.customer.domain.Customer;
-import woowacourse.shoppingcart.exception.badrequest.InvalidLoginException;
-import woowacourse.shoppingcart.exception.notfound.NotFoundCustomerException;
+import woowacourse.shoppingcart.auth.exception.badrequest.InvalidLoginException;
+import woowacourse.shoppingcart.customer.exception.notfound.NotFoundCustomerException;
 import woowacourse.shoppingcart.support.JwtTokenProvider;
 
 @Service
