+++ src/main/java/woowacourse/shoppingcart/cart/application/CartService.java
@@ -7,9 +7,9 @@ import woowacourse.shoppingcart.cart.dao.CartItemDao;
 import woowacourse.shoppingcart.cart.domain.Cart;
 import woowacourse.shoppingcart.cart.dto.QuantityChangingRequest;
 import woowacourse.shoppingcart.customer.domain.Customer;
-import woowacourse.shoppingcart.exception.badrequest.DuplicateCartItemException;
-import woowacourse.shoppingcart.exception.badrequest.NoExistCartItemException;
-import woowacourse.shoppingcart.exception.notfound.NotFoundCartException;
+import woowacourse.shoppingcart.cart.exception.badrequest.DuplicateCartItemException;
+import woowacourse.shoppingcart.cart.exception.badrequest.NoExistCartItemException;
+import woowacourse.shoppingcart.cart.exception.notfound.NotFoundCartException;
 import woowacourse.shoppingcart.product.application.ProductService;
 
 @Service
