+++ src/main/java/woowacourse/shoppingcart/cart/exception/badrequest/DuplicateCartItemException.java
@@ -1,4 +1,6 @@
-package woowacourse.shoppingcart.exception.badrequest;
+package woowacourse.shoppingcart.cart.exception.badrequest;
+
+import woowacourse.shoppingcart.exception.BadRequestException;
 
 public class DuplicateCartItemException extends BadRequestException {
