+++ src/main/java/woowacourse/shoppingcart/customer/domain/Password.java
@@ -3,7 +3,7 @@ package woowacourse.shoppingcart.customer.domain;
 import java.util.Objects;
 import java.util.regex.Pattern;
 import org.springframework.security.crypto.bcrypt.BCrypt;
-import woowacourse.shoppingcart.exception.badrequest.InvalidPasswordException;
+import woowacourse.shoppingcart.customer.exception.badrequest.InvalidPasswordException;
 
 public class Password {
