+++ src/test/java/woowacourse/shoppingcart/unit/auth/application/AuthServiceTest.java
@@ -12,8 +12,8 @@ import org.springframework.security.crypto.bcrypt.BCrypt;
 import woowacourse.shoppingcart.auth.application.AuthService;
 import woowacourse.shoppingcart.customer.domain.Customer;
 import woowacourse.shoppingcart.auth.dto.LoginRequest;
-import woowacourse.shoppingcart.exception.badrequest.InvalidLoginException;
-import woowacourse.shoppingcart.exception.notfound.NotFoundCustomerException;
+import woowacourse.shoppingcart.auth.exception.badrequest.InvalidLoginException;
+import woowacourse.shoppingcart.customer.exception.notfound.NotFoundCustomerException;
 import woowacourse.shoppingcart.unit.ServiceMockTest;
 
 class AuthServiceTest extends ServiceMockTest {
