+++ src/test/java/woowacourse/shoppingcart/unit/cart/application/CartServiceTest.java
@@ -14,10 +14,10 @@ import woowacourse.shoppingcart.cart.dao.CartItemDao;
 import woowacourse.shoppingcart.cart.domain.Cart;
 import woowacourse.shoppingcart.cart.dto.QuantityChangingRequest;
 import woowacourse.shoppingcart.customer.domain.Customer;
-import woowacourse.shoppingcart.exception.badrequest.DuplicateCartItemException;
-import woowacourse.shoppingcart.exception.badrequest.NoExistCartItemException;
-import woowacourse.shoppingcart.exception.notfound.NotFoundCartException;
-import woowacourse.shoppingcart.exception.notfound.NotFoundProductException;
+import woowacourse.shoppingcart.cart.exception.badrequest.DuplicateCartItemException;
+import woowacourse.shoppingcart.cart.exception.badrequest.NoExistCartItemException;
+import woowacourse.shoppingcart.cart.exception.notfound.NotFoundCartException;
+import woowacourse.shoppingcart.product.exception.notfound.NotFoundProductException;
 import woowacourse.shoppingcart.product.domain.Product;
 import woowacourse.shoppingcart.unit.ServiceMockTest;
