+++ src/test/java/woowacourse/shoppingcart/unit/cart/ui/CartItemControllerTest.java
@@ -30,9 +30,9 @@ import woowacourse.shoppingcart.cart.domain.Cart;
 import woowacourse.shoppingcart.cart.dto.CartItemAdditionRequest;
 import woowacourse.shoppingcart.cart.dto.QuantityChangingRequest;
 import woowacourse.shoppingcart.customer.domain.Customer;
-import woowacourse.shoppingcart.exception.badrequest.DuplicateCartItemException;
-import woowacourse.shoppingcart.exception.badrequest.NoExistCartItemException;
-import woowacourse.shoppingcart.exception.notfound.NotFoundProductException;
+import woowacourse.shoppingcart.cart.exception.badrequest.DuplicateCartItemException;
+import woowacourse.shoppingcart.cart.exception.badrequest.NoExistCartItemException;
+import woowacourse.shoppingcart.product.exception.notfound.NotFoundProductException;
 import woowacourse.shoppingcart.product.domain.Product;
 import woowacourse.shoppingcart.unit.ControllerTest;
