+++ src/test/java/woowacourse/shoppingcart/unit/customer/application/CustomerServiceTest.java
@@ -13,8 +13,8 @@ import woowacourse.shoppingcart.customer.application.CustomerService;
 import woowacourse.shoppingcart.customer.domain.Customer;
 import woowacourse.shoppingcart.customer.dto.CustomerCreationRequest;
 import woowacourse.shoppingcart.customer.dto.CustomerUpdationRequest;
-import woowacourse.shoppingcart.exception.badrequest.DuplicateEmailException;
-import woowacourse.shoppingcart.exception.notfound.NotFoundCustomerException;
+import woowacourse.shoppingcart.customer.exception.badrequest.DuplicateEmailException;
+import woowacourse.shoppingcart.customer.exception.notfound.NotFoundCustomerException;
 import woowacourse.shoppingcart.unit.ServiceMockTest;
 
 class CustomerServiceTest extends ServiceMockTest {
