+++ src/test/java/woowacourse/shoppingcart/unit/customer/domain/EmailTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import woowacourse.shoppingcart.customer.domain.Email;
-import woowacourse.shoppingcart.exception.badrequest.InvalidEmailException;
+import woowacourse.shoppingcart.customer.exception.badrequest.InvalidEmailException;
 
 class EmailTest {
