+++ src/test/java/woowacourse/shoppingcart/unit/product/ui/ProductControllerTest.java
@@ -18,7 +18,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.ResultActions;
-import woowacourse.shoppingcart.exception.notfound.NotFoundProductException;
+import woowacourse.shoppingcart.product.exception.notfound.NotFoundProductException;
 import woowacourse.shoppingcart.product.domain.Product;
 import woowacourse.shoppingcart.product.dto.ProductsResponse;
 import woowacourse.shoppingcart.unit.ControllerTest;
